+++ R/baRulho-package.R
@@ -32,9 +32,5 @@
 #'
 #' @docType package
 #' @name baRulho
-#' @keywords internal
-"_PACKAGE"
 #' @details License: GPL (>= 2)
-NULL
-#> NULL
-#'
+"_PACKAGE"