+++ tests/testthat/test-plot_aligned_sounds.R
@@ -2,9 +2,8 @@ test_that("basic", {
     # load example data
     data("test_sounds_est")
 
-    
-    
-
+  unlink(list.files(tempdir(), pattern = "jpeg$"))
+  
   # plot (look into temporary working directory `tempdir()`)
   plot_aligned_sounds(X = test_sounds_est[test_sounds_est$sound.files == test_sounds_est$sound.files[1]], dest.path = tempdir(), duration = 1, ovlp = 0)